refactor(service): move subjectmappingbuiltin back to service/internal/
Move subjectmappingbuiltin from service/pkg/ back to service/internal/
now that no cross-module imports remain. The claims_attributes_test.go
unit test (selector evaluation logic) moves into the package itself.
The keycloak_attributes_test.go integration test is refactored to assert
on flattened entity keys instead of calling EvaluateSubjectMappings
directly.

- service/pkg/subjectmappingbuiltin/ → service/internal/subjectmappingbuiltin/
- claims_attributes_test.go → service/internal/subjectmappingbuiltin/
- keycloak_attributes_test.go refactored to remove subjectmappingbuiltin import
- Update all 8 import paths within service/
