+++ service/internal/access/v2/evaluate.go
@@ -11,8 +11,8 @@ import (
 	authz "github.com/opentdf/platform/protocol/go/authorization/v2"
 	"github.com/opentdf/platform/protocol/go/policy"
 	attrs "github.com/opentdf/platform/protocol/go/policy/attributes"
+	"github.com/opentdf/platform/service/internal/subjectmappingbuiltin"
 	"github.com/opentdf/platform/service/logger"
-	"github.com/opentdf/platform/service/pkg/subjectmappingbuiltin"
 )
 
 var (
