+++ service/internal/access/v2/evaluate_test.go
@@ -9,8 +9,8 @@ import (
 	authz "github.com/opentdf/platform/protocol/go/authorization/v2"
 	"github.com/opentdf/platform/protocol/go/policy"
 	attrs "github.com/opentdf/platform/protocol/go/policy/attributes"
+	"github.com/opentdf/platform/service/internal/subjectmappingbuiltin"
 	"github.com/opentdf/platform/service/logger"
-	"github.com/opentdf/platform/service/pkg/subjectmappingbuiltin"
 	"github.com/opentdf/platform/service/policy/actions"
 )
