+++ service/internal/subjectmappingbuiltin/subject_mapping_builtin_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/opentdf/platform/protocol/go/entityresolution"
 	"github.com/opentdf/platform/protocol/go/policy"
 	"github.com/opentdf/platform/protocol/go/policy/attributes"
-	"github.com/opentdf/platform/service/pkg/subjectmappingbuiltin"
+	"github.com/opentdf/platform/service/internal/subjectmappingbuiltin"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/types/known/structpb"
