+++ test/integration/entityresolution/keycloak_attributes_test.go
@@ -9,25 +9,22 @@ import (
 
 	"github.com/Nerzal/gocloak/v13"
 	"github.com/opentdf/platform/lib/flattening"
-	"github.com/opentdf/platform/protocol/go/entityresolution"
-	"github.com/opentdf/platform/protocol/go/policy"
-	"github.com/opentdf/platform/protocol/go/policy/attributes"
-	"github.com/opentdf/platform/service/pkg/subjectmappingbuiltin"
 	"github.com/opentdf/platform/test/integration/entityresolution/internal"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/protobuf/types/known/structpb"
 )
 
 // TestKeycloakUserAttributeSubjectMapping verifies that when using the Keycloak ERS,
-// subject mapping selectors must match the Keycloak user object structure, NOT raw JWT
-// claim names. Custom Keycloak user attributes are nested under `.attributes.<name>[]`,
-// so a selector like `.department` will never match even if the JWT contains
-// `"department": "Finance"`. The correct selector is `.attributes.department[]`.
+// custom user attributes are nested under ".attributes.<name>[]" in the flattened
+// entity, NOT at the top level as ".<name>". This means subject mapping selectors
+// must use ".attributes.department[]" rather than ".department".
 //
 // This test exists to document and verify actual Keycloak ERS behavior, and to serve
 // as a reference for community members debugging subject mapping failures.
 // See: https://github.com/orgs/opentdf/discussions/3115
+//
+// The actual selector evaluation logic is unit-tested in
+// service/internal/subjectmappingbuiltin/claims_attributes_test.go.
 func TestKeycloakUserAttributeSubjectMapping(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skipping Keycloak integration tests in short mode")
@@ -82,72 +79,27 @@ func TestKeycloakUserAttributeSubjectMapping(t *testing.T) {
 	err = json.Unmarshal(userJSON, &genericMap)
 	require.NoError(t, err)
 
-	entityStruct, err := structpb.NewStruct(genericMap)
-	require.NoError(t, err)
-
-	entityRepresentation := &entityresolution.EntityRepresentation{
-		OriginalId:      "jen",
-		AdditionalProps: []*structpb.Struct{entityStruct},
-	}
-
-	// Log all flattened keys so the structure is visible in test output
+	// Flatten the entity and verify the key structure matches what the
+	// subject mapping evaluator will see at runtime.
 	flattenedEntity, err := flattening.Flatten(genericMap)
 	require.NoError(t, err)
-	t.Log("Flattened Keycloak user keys:")
+
+	flatKeys := make(map[string]interface{})
 	for _, item := range flattenedEntity.Items {
 		t.Logf("  key=%q value=%v", item.Key, item.Value)
+		flatKeys[item.Key] = item.Value
 	}
 
-	const attrFQN = "https://example.com/attr/department/value/finance"
-
-	t.Run("correct selector matches Keycloak user attribute", func(t *testing.T) {
-		entitlements, err := subjectmappingbuiltin.EvaluateSubjectMappings(
-			buildAttributeSubjectMapping(attrFQN, ".attributes.department[]", "Finance"),
-			entityRepresentation,
-		)
-		require.NoError(t, err)
-		assert.Equal(t, []string{attrFQN}, entitlements,
-			"selector '.attributes.department[]' should match Keycloak user attribute")
+	t.Run("correct selector key exists in flattened entity", func(t *testing.T) {
+		// The Keycloak ERS nests custom attributes under .attributes.<name>[]
+		assert.Contains(t, flatKeys, ".attributes.department[]",
+			"Keycloak user attribute should be flattened as '.attributes.department[]'")
 	})
 
-	t.Run("JWT claim name selector does not match Keycloak user object", func(t *testing.T) {
-		entitlements, err := subjectmappingbuiltin.EvaluateSubjectMappings(
-			buildAttributeSubjectMapping(attrFQN, ".department", "Finance"),
-			entityRepresentation,
-		)
-		require.NoError(t, err)
-		assert.Empty(t, entitlements,
-			"selector '.department' should NOT match: Keycloak ERS resolves user objects, not JWT claims")
+	t.Run("JWT claim name key does not exist in flattened entity", func(t *testing.T) {
+		// A bare .department key should NOT appear — this is why selectors
+		// like ".department" fail against Keycloak ERS entities.
+		assert.NotContains(t, flatKeys, ".department",
+			"bare '.department' should NOT be a key in a Keycloak user entity")
 	})
-}
-
-func buildAttributeSubjectMapping(attrFQN, selector, value string) map[string]*attributes.GetAttributeValuesByFqnsResponse_AttributeAndValue { //nolint:unparam // attrFQN is parameterized so callers can specify any attribute FQN
-	return map[string]*attributes.GetAttributeValuesByFqnsResponse_AttributeAndValue{
-		attrFQN: {
-			Value: &policy.Value{
-				SubjectMappings: []*policy.SubjectMapping{
-					{
-						SubjectConditionSet: &policy.SubjectConditionSet{
-							SubjectSets: []*policy.SubjectSet{
-								{
-									ConditionGroups: []*policy.ConditionGroup{
-										{
-											BooleanOperator: policy.ConditionBooleanTypeEnum_CONDITION_BOOLEAN_TYPE_ENUM_AND,
-											Conditions: []*policy.Condition{
-												{
-													SubjectExternalSelectorValue: selector,
-													Operator:                     policy.SubjectMappingOperatorEnum_SUBJECT_MAPPING_OPERATOR_ENUM_IN,
-													SubjectExternalValues:        []string{value},
-												},
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
 }
